fixing the build error
