+++ src/app/bts-fifa-2026/constants.ts
@@ -1,0 +1,18 @@
+export const albums = [
+  "2 Cool 4 Skool",
+  "O!RUL8,2?",
+  "Skool Luv Affair",
+  "Dark & Wild",
+  "The Most Beautiful Moment in Life, Pt. 1",
+  "The Most Beautiful Moment in Life, Pt. 2",
+  "The Most Beautiful Moment in Life: Young Forever",
+  "Wings",
+  "You Never Walk Alone",
+  "Love Yourself: Tear",
+  "Love Yourself: Answer",
+  "Map of the Soul: Persona",
+  "Map of the Soul: 7",
+  "BE",
+  "Proof",
+  "Arirang",
+];
